always render Competition pages from the Server
