+++ next-frontend/src/app/competitions/[competitionId]/admin/page.tsx
@@ -1,0 +1,32 @@
+import { Heading } from '@chakra-ui/react'
+import { Container } from '@chakra-ui/react'
+import Link from "next/link";
+import {getCompetitionInfo} from "@/lib/wca/competitions/getCompetitionInfo";
+import PermissionCheck from "@/components/PermissionCheck";
+
+export default async function CompetitionOverview({ params }: { params: Promise<{ competitionId: string }> }){
+  const { competitionId } = await params;
+  const { data: competitionInfo, error } = await getCompetitionInfo(competitionId);
+
+  if(error){
+    return <p>
+      Error fetching competition
+    </p>
+  }
+
+  if(!competitionInfo){
+    return <p>
+      Competition does not exist
+    </p>
+  }
+
+  return (
+    <Container centerContent>
+      <Heading>{competitionInfo.id}</Heading>
+      <PermissionCheck permissionCheck={(permission) => permission.canAdministerCompetition(competitionId)}>
+        <p>You are administering this competition</p>
+        <p>Go back to the public page <Link href={`/competitions/${competitionInfo.id}`}>here</Link></p>
+      </PermissionCheck>
+    </Container>
+  )
+}
