+++ next-frontend/src/app/competitions/[competitionId]/page.tsx
@@ -1,11 +1,11 @@
 import { Container, Heading } from "@chakra-ui/react";
 import PermissionProvider from "@/providers/PermissionProvider";
 import PermissionsTestMessage from "@/components/competitions/permissionsTestMessage";
-import { serverClient } from "@/lib/wca/wcaAPI";
+import { getCompetitionInfo } from "@/lib/wca/competitions/getCompetitionInfo";
 
 export default async function CompetitionOverView({ params }: { params: Promise<{ competitionId: string }> }){
   const { competitionId } = await params;
-  const { data: competitionInfo, error } = await serverClient.GET("/competitions/{competitionId}/", { params: { path: { competitionId }}})
+  const { data: competitionInfo, error } = await getCompetitionInfo(competitionId);
 
   if(error){
     return <p>
