+++ next-frontend/src/lib/wca/competitions/getCompetitionInfo.tsx
@@ -1,0 +1,6 @@
+import { serverClient } from "@/lib/wca/wcaAPI";
+import { cache } from "react";
+
+export const getCompetitionInfo = cache(async (competitionId: string) => {
+  return await serverClient.GET("/competitions/{competitionId}/", { params: { path: { competitionId }}})
+});
